Fix: Unify uniforms
