+++ addons/gsri_equipment/Data/uniforms/gsri_shirt_camo_co.paa
