+++ addons/gsri_equipment/Data/uniforms/gsri_shirt_green_co.paa
