+++ addons/gsri_equipment/Data/uniforms/gsri_uniform.rvmat
