+++ addons/gsri_equipment/Data/uniforms/gsri_uniform_co.paa
