--- addons/gsri_uniforms/$PBOPREFIX$
@@ -1,1 +1,0 @@
-fr\gsri\uniforms
